:hammer: refactor: 중복된 controller, service 합침
