+++ gg-recruit-api/src/main/java/gg/recruit/api/admin/controller/RecruitmentAdminController.java
@@ -4,6 +4,8 @@ import javax.validation.Valid;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,7 +13,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 import gg.data.recruit.recruitment.Recruitment;
 import gg.recruit.api.admin.controller.request.RecruitmentCreateReqDto;
+import gg.recruit.api.admin.controller.request.UpdateStatusRequestDto;
 import gg.recruit.api.admin.service.RecruitmentAdminService;
+import gg.recruit.api.admin.service.dto.UpdateRecruitStatusParam;
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -25,5 +29,14 @@ public class RecruitmentAdminController {
 		Recruitment recruitment = recruitmentDto.toRecruitment();
 		recruitmentAdminService.createRecruitment(recruitment, recruitmentDto.getForm());
 		return ResponseEntity.status(HttpStatus.CREATED).build();
+	}
+
+	@PatchMapping("/{recruitId}/status")
+	public ResponseEntity<Void> updateRecruitStatus(@PathVariable Long recruitId,
+		@RequestBody UpdateStatusRequestDto requestDto) {
+
+		recruitmentAdminService.updateRecruitStatus(
+			new UpdateRecruitStatusParam(recruitId, requestDto.getFinish()));
+		return ResponseEntity.noContent().build();
 	}
 }
